data/reddit/requirements.txt
# Python packages for Reddit scraping
praw>=7.7.1
pandas>=2.0.0
python-dotenv>=1.0.0
